# triemux: make shouldRedirToLowercase more efficient by regex compilation+reuse