@sengi 2024-04-03
Just to let you know this is in production now. Thanks for contributing!

![CPU usage graph showing similar usage before and after the rollout, perhaps a slight reduction](https://github.com/alphagov/router/assets/1170635/500de8f6-5ff1-4f89-a65a-a32e111f4424)